model/code/requirements.txt
torch
pillow

opencv-python
facenet-pytorch

git+https://github.com/meilof/pysnark
